chore: cleanup and docs
