+++ src/cyclefold/mod.rs
@@ -1,8 +1,9 @@
-//! This module defines Cyclefold folding scheme and its related functions.
+//! This module defines CycleFold folding scheme and its related functions.
 
 mod circuit;
 mod gadgets;
 mod nova_circuit;
+mod util;
 
 pub mod nifs;
 pub mod snark;
