+++ src/cyclefold/nifs.rs
@@ -2,65 +2,16 @@
 
 use std::marker::PhantomData;
 
-use ff::Field;
-
 use crate::{
-  constants::{
-    BN_LIMB_WIDTH, BN_N_LIMBS, NIO_CYCLE_FOLD, NUM_CHALLENGE_BITS, NUM_FE_IN_EMULATED_POINT,
-  },
+  constants::{NIO_CYCLE_FOLD, NUM_CHALLENGE_BITS, NUM_FE_IN_EMULATED_POINT},
   errors::NovaError,
-  gadgets::{f_to_nat, nat_to_limbs, scalar_as_base},
+  gadgets::scalar_as_base,
   r1cs::{R1CSInstance, R1CSShape, R1CSWitness, RelaxedR1CSInstance, RelaxedR1CSWitness},
   traits::{commitment::CommitmentTrait, AbsorbInROTrait, Engine, ROConstants, ROTrait},
   CommitmentKey, CompressedCommitment,
 };
 
-/// Absorb a commitment over engine `E1` into an RO over engine `E2` by absorbing the limbs
-pub fn absorb_primary_commitment<E1, E2>(
-  comm: &impl CommitmentTrait<E1>,
-  ro: &mut impl ROTrait<E2::Base, E2::Scalar>,
-) where
-  E1: Engine<Base = <E2 as Engine>::Scalar>,
-  E2: Engine<Base = <E1 as Engine>::Scalar>,
-{
-  let (x, y, is_infinity) = comm.to_coordinates();
-
-  let x_limbs = nat_to_limbs(&f_to_nat(&x), BN_LIMB_WIDTH, BN_N_LIMBS).unwrap();
-  let y_limbs = nat_to_limbs(&f_to_nat(&y), BN_LIMB_WIDTH, BN_N_LIMBS).unwrap();
-
-  for limb in x_limbs {
-    ro.absorb(scalar_as_base::<E2>(limb));
-  }
-  for limb in y_limbs {
-    ro.absorb(scalar_as_base::<E2>(limb));
-  }
-  if is_infinity {
-    ro.absorb(<E1 as Engine>::Scalar::ONE);
-  } else {
-    ro.absorb(<E1 as Engine>::Scalar::ZERO);
-  }
-}
-
-fn absorb_primary_r1cs<E1, E2>(u: &R1CSInstance<E1>, ro: &mut impl ROTrait<E2::Base, E2::Scalar>)
-where
-  E1: Engine<Base = <E2 as Engine>::Scalar>,
-  E2: Engine<Base = <E1 as Engine>::Scalar>,
-{
-  absorb_primary_commitment::<E1, E2>(&u.comm_W, ro);
-  for x in &u.X {
-    ro.absorb(*x);
-  }
-}
-
-fn absorb_cyclefold_r1cs<E: Engine>(u: &R1CSInstance<E>, ro: &mut E::RO) {
-  u.comm_W.absorb_in_ro(ro);
-  u.X.iter().for_each(|x| {
-    let limbs: Vec<E::Scalar> = nat_to_limbs(&f_to_nat(x), BN_LIMB_WIDTH, BN_N_LIMBS).unwrap();
-    limbs
-      .into_iter()
-      .for_each(|limb| ro.absorb(scalar_as_base::<E>(limb)));
-  });
-}
+use super::util::{absorb_cyclefold_r1cs, absorb_primary_commitment, absorb_primary_r1cs};
 
 /// A SNARK that holds the proof of a step of an incremental computation of the primary circuit
 /// in the CycleFold folding scheme.
